requirements.txt
flask==3.1.2
diffusers==0.36.0
torch==2.9.0
transformers==4.57.1
pillow==12.0.0
numpy==2.3.3
accelerate==0.34.2